Added authentication based on TU Delft LDAP

+++ src/main/java/nl/tudelft/ewi/devhub/server/web/resources/BuildServerResource.java
@@ -21,71 +21,83 @@ import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 
 import nl.tudelft.ewi.devhub.server.backend.BuildsBackend;
-import nl.tudelft.ewi.devhub.server.database.controllers.Users;
 import nl.tudelft.ewi.devhub.server.database.entities.BuildServer;
-import nl.tudelft.ewi.devhub.server.database.entities.User;
 import nl.tudelft.ewi.devhub.server.web.errors.ApiError;
+import nl.tudelft.ewi.devhub.server.web.filters.RequestScope;
+import nl.tudelft.ewi.devhub.server.web.filters.RequireAuthenticatedUser;
 import nl.tudelft.ewi.devhub.server.web.templating.TemplateEngine;
 
 import org.eclipse.jetty.util.UrlEncoded;
+import org.jboss.resteasy.plugins.guice.RequestScoped;
 
 import com.google.common.base.Strings;
 import com.google.common.collect.Maps;
 import com.google.inject.persist.Transactional;
 
+@RequestScoped
 @Path("build-servers")
 @Produces(MediaType.TEXT_HTML)
+@RequireAuthenticatedUser
 public class BuildServerResource {
 
-	private static final long USER_ID = 1;
-	
 	private final TemplateEngine templateEngine;
 	private final BuildsBackend backend;
-	private final Users users;
+	private final RequestScope scope;
 
 	@Inject
-	BuildServerResource(TemplateEngine templateEngine, BuildsBackend backend, Users users) {
+	BuildServerResource(TemplateEngine templateEngine, BuildsBackend backend, RequestScope scope) {
 		this.templateEngine = templateEngine;
 		this.backend = backend;
-		this.users = users;
+		this.scope = scope;
 	}
 
 	@GET
-	public String showUserPage(@Context HttpServletRequest request, @QueryParam("error") String error) throws IOException {
-		User requester = users.find(USER_ID);
-
+	public Response showBuildServers(@Context HttpServletRequest request, @QueryParam("error") String error) 
+			throws IOException, URISyntaxException {
+		
+		if (!scope.isAdmin()) {
+			return Response.seeOther(new URI("/")).build();
+		}
+		
 		Map<String, Object> parameters = Maps.newHashMap();
-		parameters.put("user", requester);
+		parameters.put("user", scope.getUser());
 		parameters.put("servers", backend.listActiveBuildServers());
 		if (!Strings.isNullOrEmpty(error)) {
 			parameters.put("error", error);
 		}
 
 		List<Locale> locales = Collections.list(request.getLocales());
-		return templateEngine.process("build-servers.ftl", locales, parameters);
+		return Response.ok(templateEngine.process("build-servers.ftl", locales, parameters)).build();
 	}
 	
 	@GET
 	@Path("setup")
-	public String showNewBuildServerSetupPage(@Context HttpServletRequest request, @QueryParam("error") String error) 
-			throws IOException {
+	public Response showNewBuildServerSetupPage(@Context HttpServletRequest request, @QueryParam("error") String error) 
+			throws IOException, URISyntaxException {
 		
-		User requester = users.find(USER_ID);
+		if (!scope.isAdmin()) {
+			return Response.seeOther(new URI("/")).build();
+		}
+		
 		List<Locale> locales = Collections.list(request.getLocales());
 		
 		Map<String, Object> parameters = Maps.newHashMap();
-		parameters.put("user", requester);
+		parameters.put("user", scope.getUser());
 		if (!Strings.isNullOrEmpty(error)) {
 			parameters.put("error", error);
 		}
 		
-		return templateEngine.process("build-server-setup.ftl", locales, parameters);
+		return Response.ok(templateEngine.process("build-server-setup.ftl", locales, parameters)).build();
 	}
 	
 	@POST
 	@Path("setup")
 	public Response addNewBuildServer(@FormParam("name") String name, @FormParam("secret") String secret, @FormParam("host") String host) 
 			throws URISyntaxException {
+		
+		if (!scope.isAdmin()) {
+			return Response.seeOther(new URI("/")).build();
+		}
 		
 		try {
 			BuildServer server = new BuildServer();
@@ -106,6 +118,10 @@ public class BuildServerResource {
 	@Path("delete")
 	@Transactional
 	public Response deleteBuildServer(@FormParam("id") long id) throws URISyntaxException {
+		if (!scope.isAdmin()) {
+			return Response.seeOther(new URI("/")).build();
+		}
+		
 		try {
 			backend.deleteBuildServer(id);
 			return Response.seeOther(new URI("/build-servers")).build();
